Remove the idependent prefix selector component
